+++ packages/create-moralis-dapp/src/generators/next/template/app/components/templates/index.ts
@@ -1,0 +1,2 @@
+export * from './UserPage';
+export * from './AuthPage';