+++ packages/next/src/hooks/auth/useAuthRequestChallengeSolana.ts
@@ -1,12 +1,9 @@
 import {
   requestChallengeSolanaOperation as operation,
   RequestChallengeSolanaRequest,
-  RequestChallengeSolanaResponse,
 } from '@moralisweb3/common-auth-utils';
-import { fetcher } from '../../utils/fetcher';
 import { FetchParams } from '../types';
-import { useCallback } from 'react';
-import useSWR from 'swr';
+import { useResolver } from '../resolvers';
 
 export type RequestChallengeSolanaRequestClient = Pick<RequestChallengeSolanaRequest, 'address' | 'network'>;
 
@@ -14,39 +11,25 @@ export const useAuthRequestChallengeSolana = (
   request?: RequestChallengeSolanaRequestClient,
   fetchParams?: FetchParams,
 ) => {
-  const endpoint = 'auth/requestChallengeSolana';
-  const { deserializeResponse } = operation;
-
-  const { data, error, isValidating, mutate } = useSWR<RequestChallengeSolanaResponse>(
-    [
-      endpoint,
-      {
-        deserializeResponse,
-        request,
-      },
-    ],
-    request ? fetcher : null,
-    {
-      revalidateOnFocus: false,
-      revalidateIfStale: false,
-      ...fetchParams,
-    },
-  );
-
-  const requestChallengeAsync = useCallback((params: RequestChallengeSolanaRequestClient) => {
-    return mutate(
-      fetcher(endpoint, {
-        deserializeResponse,
-        request: params,
-      }),
-    );
-  }, []);
+  const { data, error, fetch, isFetching } = useResolver({
+    endpoint: 'auth/requestChallengeSolana',
+    operation,
+    request,
+    fetchParams,
+  });
 
   return {
     challenge: data,
     error,
-    requestChallengeAsync,
-    refetch: async () => mutate(),
-    isValidating,
+    requestChallengeAsync: fetch,
+    /**
+     * @deprecated use `fetch()` instead
+     */
+    refetch: () => fetch(),
+    isFetching,
+    /**
+     * @deprecated use `isFetching` instead
+     */
+    isValidating: isFetching,
   };
 };