+++ packages/client/firebaseEvmAuth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moralisweb3/client-firebase-evm-auth
 
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-firebase-auth-utils@2.9.0
+
+## 2.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-firebase-auth-utils@2.8.2
+
 ## 2.8.1
 
 ### Patch Changes